Revert "added list with checkboxes"
This reverts commit de01db4d8cc59580ffffcccf8f454837355f4022.
